# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Dokumentasi SIMRS - Kelompok 2",
        "description": "Dokumentasi API untuk Modul Pemeriksaan (Perawat & Dokter)",
        "contact": {
            "email": "hipnihaja@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "API Server Lokal"
        }
    ],
    "paths": {
        "/asesmen": {
            "post": {
                "tags": [
                    "Asesmen Perawat"
                ],
                "summary": "Sipman data asesmen awal (TTV)",
                "description": "Menyimpan data keluhan, tensi, suhu yang diinput oleh perawat.",
                "operationId": "storeAsesmen",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_antrian",
                                    "id_pasien",
                                    "keluhan_utama",
                                    "tensi",
                                    "suhu",
                                    "nadi",
                                    "respirasi",
                                    "tinggi_badan",
                                    "berat_badan"
                                ],
                                "properties": {
                                    "id_antrian": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_pasien": {
                                        "type": "integer",
                                        "example": 99
                                    },
                                    "keluhan_utama": {
                                        "type": "string",
                                        "example": "Nyeri dada"
                                    },
                                    "tensi": {
                                        "type": "string",
                                        "example": "120/80"
                                    },
                                    "suhu": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 36.4
                                    },
                                    "nadi": {
                                        "type": "integer",
                                        "example": 80
                                    },
                                    "respirasi": {
                                        "type": "integer",
                                        "example": 20
                                    },
                                    "tinggi_badan": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 165.6
                                    },
                                    "berat_badan": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 55.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil menyimpan data"
                    },
                    "422": {
                        "description": "Validasi Error"
                    }
                }
            }
        },
        "/rekam-medik": {
            "post": {
                "tags": [
                    "Rekam Medik Dokter"
                ],
                "summary": "Simpan data rekam medik (Diagnosa & Tindakan)",
                "description": "Menyimpan hasil pemeriksaan, diagnosa, dan tindakan medis yang diinput oleh dokter.",
                "operationId": "storeRekamMedik",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_antrian",
                                    "id_pasien",
                                    "id_dokter",
                                    "diagnosa",
                                    "tindakan_medis"
                                ],
                                "properties": {
                                    "id_antrian": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_pasien": {
                                        "type": "integer",
                                        "example": 99
                                    },
                                    "id_dokter": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "diagnosa": {
                                        "type": "string",
                                        "example": "Hipertensi Grade II"
                                    },
                                    "tindakan_medis": {
                                        "type": "string",
                                        "example": "Pemeriksaan EKG + Konsultasi Gizi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil menyimpan data rekam medik",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Rekam medik pasien berhasil disimpan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi Error"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Asesmen Perawat",
            "description": "Asesmen Perawat"
        },
        {
            "name": "Rekam Medik Dokter",
            "description": "Rekam Medik Dokter"
        }
    ]
}